refactor: cleanup data loading (#752)
* refactor: cleanup data loading

* refactor: bump version

* fix: import

* fix: cyclic import
