--- now/app/base/flow.yml
@@ -1,16 +1,0 @@
-jtype: Flow
-with:
-  name: nowapi
-  monitoring: true
-  port_expose: 8080
-  prefetch: 10
-  cors: true
-  env:
-    JINA_LOG_LEVEL: DEBUG
-jcloud:
-  version: ${{ ENV.JINA_VERSION }}
-  custom_dns:
-    - ${{ ENV.CUSTOM_DNS }}
-  labels:
-    team: now
-executors:
