--- now/app/base/playground.py
@@ -1,1 +1,0 @@
-# TODO custom playground in the future possible
