+++ now/data_loading/data_loading.py
@@ -1,22 +1,26 @@
+import base64
 import os
+import pathlib
+import pickle
 import uuid
 from collections import defaultdict
 from copy import deepcopy
+from os.path import join as osp
 
 from docarray import Document, DocumentArray
 
 from now.app.base.app import JinaNOWApp
-from now.constants import DatasetTypes
-from now.data_loading.es import ElasticsearchExtractor
-from now.data_loading.utils import (
-    fetch_da_from_url,
-    get_dataset_url,
-    get_s3_bucket_and_folder_prefix,
+from now.constants import (
+    BASE_STORAGE_URL,
+    DEMO_DATASET_DOCARRAY_VERSION,
+    DatasetTypes,
+    Modalities,
 )
+from now.data_loading.elasticsearch import ElasticsearchExtractor
 from now.demo_data import DemoDatasetNames
 from now.log import yaspin_extended
 from now.now_dataclasses import UserInput
-from now.utils import sigmap
+from now.utils import download, sigmap
 
 
 def load_data(app: JinaNOWApp, user_input: UserInput) -> DocumentArray:
@@ -186,6 +190,74 @@ def _list_files_from_s3_bucket(app: JinaNOWApp, user_input: UserInput) -> Docume
                         break
 
     return DocumentArray(docs)
+
+
+def fetch_da_from_url(
+    url: str, downloaded_path: str = '~/.cache/jina-now'
+) -> DocumentArray:
+    data_dir = os.path.expanduser(downloaded_path)
+    if not os.path.exists(osp(data_dir, 'data/tmp')):
+        os.makedirs(osp(data_dir, 'data/tmp'))
+    data_path = (
+        data_dir
+        + f"/data/tmp/{base64.b64encode(bytes(url, 'utf-8')).decode('utf-8')}.bin"
+    )
+    if not os.path.exists(data_path):
+        download(url, data_path)
+
+    try:
+        da = DocumentArray.load_binary(data_path)
+    except pickle.UnpicklingError:
+        path = pathlib.Path(data_path).expanduser().resolve()
+        os.remove(path)
+        download(url, data_path)
+        da = DocumentArray.load_binary(data_path)
+    return da
+
+
+def get_dataset_url(dataset: str, output_modality: Modalities) -> str:
+    data_folder = None
+    docarray_version = DEMO_DATASET_DOCARRAY_VERSION
+    if output_modality == Modalities.IMAGE:
+        data_folder = 'jpeg'
+    elif output_modality == Modalities.TEXT:
+        data_folder = 'text'
+    elif output_modality == Modalities.MUSIC:
+        data_folder = 'music'
+    elif output_modality == Modalities.VIDEO:
+        data_folder = 'video'
+    elif output_modality == Modalities.TEXT_AND_IMAGE:
+        data_folder = 'text-image'
+    if output_modality not in [
+        Modalities.MUSIC,
+        Modalities.VIDEO,
+        Modalities.TEXT_AND_IMAGE,
+    ]:
+        model_name = 'ViT-B32'
+        return f'{BASE_STORAGE_URL}/{data_folder}/{dataset}.{model_name}-{docarray_version}.bin'
+    else:
+        return f'{BASE_STORAGE_URL}/{data_folder}/{dataset}-{docarray_version}.bin'
+
+
+def get_s3_bucket_and_folder_prefix(user_input: UserInput):
+    import boto3.session
+
+    s3_uri = user_input.dataset_path
+    if not s3_uri.startswith('s3://'):
+        raise ValueError(
+            f"Can't process S3 URI {s3_uri} as it assumes it starts with: 's3://'"
+        )
+
+    bucket = s3_uri.split('/')[2]
+    folder_prefix = '/'.join(s3_uri.split('/')[3:])
+
+    session = boto3.session.Session(
+        aws_access_key_id=user_input.aws_access_key_id,
+        aws_secret_access_key=user_input.aws_secret_access_key,
+    )
+    bucket = session.resource('s3').Bucket(bucket)
+
+    return bucket, folder_prefix
 
 
 def deep_copy_da(da: DocumentArray) -> DocumentArray:
