+++ now/data_loading/elasticsearch/data_extractor.py
@@ -5,8 +5,8 @@ from typing import Dict, List, Optional, Set, Union
 from docarray import Document, DocumentArray
 from PIL import Image
 
-from now.data_loading.es.connector import ElasticsearchConnector
-from now.data_loading.transform_docarray import transform_multi_modal_data
+from now.app.base.transform_docarray import transform_multi_modal_data
+from now.data_loading.elasticsearch.connector import ElasticsearchConnector
 
 logging.getLogger("PIL.Image").setLevel(logging.CRITICAL + 1)
