--- now/data_loading/es/__init__.py
@@ -1,2 +1,0 @@
-from now.data_loading.es.connector import ElasticsearchConnector
-from now.data_loading.es.data_extractor import ElasticsearchExtractor
