--- now/data_loading/requirements.txt
@@ -1,25 +1,0 @@
---find-links https://download.pytorch.org/whl/torch_stable.html
-torch==1.10.2+cpu
-torchvision==0.11.3
-transformers==4.16.2
-finetuner==0.4.0
-docker==5.0.3
-pyfiglet==0.8.post1
-git+https://github.com/openai/CLIP.git
-pathspec==0.9.0
-matplotlib==3.5.1
-cryptography==36.0.1
-spacy==3.2.1
-streamlit==1.11.1
-jina==3.12.0
-cowsay==4.0
-portforward==0.2.7
-sklearn
-oauth2client==4.1.3
-google-api-python-client==2.38.0
-google-cloud-storage
-google-cloud-container
-yaspin==2.1.0
-prompt_toolkit==3.0.28
-protobuf==3.19.4
-pytest==7.1.1
