+++ tests/integration/data_loading/elastic/example_dataset.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional
 
 from tests.integration.data_loading.elastic.utils import create_es_index, es_insert
 
-from now.data_loading.es import ElasticsearchConnector
+from now.data_loading.elasticsearch import ElasticsearchConnector
 
 
 class ExampleDataset:
