+++ tests/unit/apps/test_preprocessing.py
@@ -4,12 +4,12 @@ import pytest
 from docarray import Document, DocumentArray
 
 from now.app.base.preprocess import preprocess_text
+from now.app.base.transform_docarray import transform_docarray
 from now.app.image_text_retrieval.app import ImageTextRetrieval
 from now.app.text_to_text_and_image.app import TextToTextAndImage
 from now.app.text_to_video.app import TextToVideo
 from now.constants import DatasetTypes
 from now.data_loading.data_loading import load_data
-from now.data_loading.transform_docarray import transform_docarray
 from now.demo_data import DemoDatasetNames
 from now.now_dataclasses import UserInput
